requirements.txt
streamlit==1.29.0
google-cloud-storage==2.12.0
pandas==2.1.3
numpy==1.24.3
torch==2.1.1
plotly==5.17.0
python-dotenv==1.0.0
Pillow==10.1.0
scikit-learn==1.3.2
luxpy==1.9.17